+++ envisage/resource/http_resource_protocol.py
@@ -36,7 +36,10 @@ class HTTPResourceProtocol(HasTraits):
         try:
             f = urlopen("http://" + address)
 
-        except HTTPError:
-            raise NoSuchResourceError("http:://" + address)
+        except HTTPError as exc:
+            # On Python 3.8 and 3.9, close() fails if the HTTPError has no fp.
+            if exc.fp is not None:
+                exc.close()
+            raise NoSuchResourceError("http://" + address)
 
         return f